Fix Issue #937: 売買判断システムの実装完了
Phase 1: リファクタリングファイルの統一
- daytrade_web_refactored.py → daytrade_web.py
- simple_test_refactored.py → simple_test.py
- real_data_provider_v2.py → real_data_provider.py
- optimized_cache_v2.py → optimized_cache.py
- test_patterns_refactored.py → test_patterns.py
- test_real_data_provider_v2.py → test_real_data_provider.py

Phase 2: 売買判断システムの実装
新機能:
- テクニカル分析エンジン (technical_indicators.py)
  * RSI, MACD, 移動平均, ボリンジャーバンド, ストキャスティクス
  * 総合的な売買シグナル生成 (BUY/SELL/HOLD)
  * リスク管理機能 (ポジションサイズ計算)

- リアルデータ統合 (recommendation_service.py)
  * Yahoo Finance APIからリアルタイム株価取得
  * 具体的なアクション提案: 何を・いつ・いくら
  * 明確な買い/売りタイミング指示

解決した問題:
✅ ランダムな推奨データ → リアルテクニカル分析
✅ 売買タイミング不明 → 明確なシグナル生成
✅ 投資金額不明 → リスク管理による適切な提案

テスト済み:
- テクニカル分析: BUYシグナル生成確認
- Webサーバー: ポート8001で動作確認
- API応答: /api/recommendations 正常動作
